v0/requirements.txt
black
pillow
pillow-heif